rename index_scan to node_by_index_scan
